+++ src/main/java/com/imb2025/calificaciones/service/jpa/CondicionFinalServiceImpl.java
@@ -39,7 +39,10 @@ public class CondicionFinalServiceImpl implements ICondicionFinalService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws Exception {
+        if (!repository.existsById(id)) {
+            throw new Exception("No se puede eliminar el id: " + id + " porque no existe");
+        }
         repository.deleteById(id);
     }
