+++ src/main/java/com/imb2025/calificaciones/service/jpa/CursadaServiceImpl.java
@@ -45,7 +45,10 @@ public class CursadaServiceImpl implements ICursadaService{
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws Exception {
+        if (!repo.existsById(id)) {
+            throw new Exception("No se puede eliminar el id: " + id + " porque no existe");
+        }
         repo.deleteById(id);
     }
