+++ src/main/java/com/imb2025/calificaciones/service/jpa/NivelMateriaServiceImpl.java
@@ -33,9 +33,11 @@ public class NivelMateriaServiceImpl implements INivelMateriaService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws Exception {
+        if (!repo.existsById(id)) {
+            throw new Exception("No se puede eliminar el id: " + id + " porque no existe");
+        }
         repo.deleteById(id);
-
     }
 
     @Override
