+++ src/main/java/com/imb2025/calificaciones/service/jpa/SedeServiceImpl.java
@@ -49,7 +49,10 @@ public class SedeServiceImpl implements ISedeService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws Exception {
+        if (!repo.existsById(id)) {
+            throw new Exception("No se puede eliminar el id: " + id + " porque no existe");
+        }
         repo.deleteById(id);
     }
